Added overall functionality to add examples and added an example for stddef.
Added a function for easy addition of examples in the CMakeLists.txt (similar to nou_add_test).

Added an example that demonstrates most of the functionality in stddef. The documentation in stddef also links to that example.

Fixed minor errors in both the comments in CMakeLists.txt and the code itself.